+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add NLUF device stub
+- Add TPSRS Somfy shutters
+
 
 ### Changed