+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add NLAS - wireless batteryless scene switch device type
-- Add BNEU, EBU, NLDD, NLAO, NLLF, NLUO, NLUP, Z3L
+- Add BNEU, EBU, NLDD, NLAO, NLLF, NLUO, NLUP, Z3L, NLTS, NLUF
 
 ### Changed
 
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Update Legrand and BTicino devices
+- Fix broken temperature setter when OTM is in the setup
 
 ### Security